优化播放端静态粒子每tick重算并修复测试编译：修改 ClientAnimationPlayer.kt 跳过静态粒子重算/复用求值缓冲，ExprBenchmark.kt 适配 VarDef 新签名

+++ src/test/kotlin/work/nekow/particledrawing/ExprBenchmark.kt
@@ -45,7 +45,7 @@ fun main() {
     // ---- 性能：快路径 5w ----
     var total = 0.0
     for (p in presets) {
-        val varDefs = p.vars.map { (name, expr) -> VarDef(name, expr, emptyList()) }
+        val varDefs = p.vars.map { (name, expr) -> VarDef(name, expr.toDouble(), emptyList()) }
         val cf = compileFunctionObject(p.code, varDefs) ?: run {
             continue
         }
@@ -68,7 +68,7 @@ fun main() {
     println()
     var ok = true
     for (p in presets) {
-        val varDefs = p.vars.map { (name, expr) -> VarDef(name, expr, emptyList()) }
+        val varDefs = p.vars.map { (name, expr) -> VarDef(name, expr.toDouble(), emptyList()) }
         val cf = compileFunctionObject(p.code, varDefs) ?: run { println("${p.name}: 回退，跳过"); continue }
         val regs = cf.allocRegs()
         val stack = cf.allocStack()
